Move primitive row ops and experimental row ops to detail
