+++ cpp/src/binaryop/compiled/struct_binary_ops.cuh
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2024, NVIDIA CORPORATION.
+ * Copyright (c) 2022-2025, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@
 #include <cudf/column/column_device_view.cuh>
 #include <cudf/column/column_view.hpp>
 #include <cudf/detail/iterator.cuh>
-#include <cudf/table/experimental/row_operators.cuh>
+#include <cudf/detail/row_operator/row_operators.cuh>
 
 #include <rmm/cuda_stream_view.hpp>
 #include <rmm/exec_policy.hpp>
@@ -52,8 +52,8 @@ struct device_comparison_functor {
   bool __device__ operator()(size_type i)
   {
     return _optional_iter[i].has_value() &&
-           _comparator(cudf::experimental::row::lhs_index_type{_is_lhs_scalar ? 0 : i},
-                       cudf::experimental::row::rhs_index_type{_is_rhs_scalar ? 0 : i});
+           _comparator(cudf::detail::row::lhs_index_type{_is_lhs_scalar ? 0 : i},
+                       cudf::detail::row::rhs_index_type{_is_rhs_scalar ? 0 : i});
   }
 
   OptionalIterator const _optional_iter;
@@ -64,7 +64,7 @@ struct device_comparison_functor {
 
 template <class BinaryOperator,
           typename PhysicalElementComparator =
-            cudf::experimental::row::lexicographic::sorting_physical_element_comparator>
+            cudf::detail::row::lexicographic::sorting_physical_element_comparator>
 void apply_struct_binary_op(mutable_column_view& out,
                             column_view const& lhs,
                             column_view const& rhs,
@@ -77,10 +77,10 @@ void apply_struct_binary_op(mutable_column_view& out,
     lhs.size(),
     is_any_v<BinaryOperator, ops::Greater, ops::GreaterEqual> ? order::DESCENDING
                                                               : order::ASCENDING);
-  auto const tlhs             = table_view{{lhs}};
-  auto const trhs             = table_view{{rhs}};
-  auto const table_comparator = cudf::experimental::row::lexicographic::two_table_comparator{
-    tlhs, trhs, compare_orders, {}, stream};
+  auto const tlhs = table_view{{lhs}};
+  auto const trhs = table_view{{rhs}};
+  auto const table_comparator =
+    cudf::detail::row::lexicographic::two_table_comparator{tlhs, trhs, compare_orders, {}, stream};
   auto outd = column_device_view::create(out, stream);
   auto optional_iter =
     cudf::detail::make_optional_iterator<bool>(*outd, nullate::DYNAMIC{out.has_nulls()});
@@ -123,8 +123,8 @@ struct struct_equality_functor {
 
   auto __device__ operator()(size_type i) const noexcept
   {
-    auto const lhs = cudf::experimental::row::lhs_index_type{_is_lhs_scalar ? 0 : i};
-    auto const rhs = cudf::experimental::row::rhs_index_type{_is_rhs_scalar ? 0 : i};
+    auto const lhs = cudf::detail::row::lhs_index_type{_is_lhs_scalar ? 0 : i};
+    auto const rhs = cudf::detail::row::rhs_index_type{_is_rhs_scalar ? 0 : i};
     return _optional_iter[i].has_value() and (_device_comparator(lhs, rhs) == _preserve_output);
   }
 
@@ -136,8 +136,8 @@ struct struct_equality_functor {
   bool _preserve_output;
 };
 
-template <typename PhysicalEqualityComparator =
-            cudf::experimental::row::equality::physical_equality_comparator>
+template <
+  typename PhysicalEqualityComparator = cudf::detail::row::equality::physical_equality_comparator>
 void apply_struct_equality_op(mutable_column_view& out,
                               column_view const& lhs,
                               column_view const& rhs,
@@ -152,10 +152,9 @@ void apply_struct_equality_op(mutable_column_view& out,
                "Unsupported operator for these types",
                cudf::data_type_error);
 
-  auto tlhs = table_view{{lhs}};
-  auto trhs = table_view{{rhs}};
-  auto table_comparator =
-    cudf::experimental::row::equality::two_table_comparator{tlhs, trhs, stream};
+  auto tlhs             = table_view{{lhs}};
+  auto trhs             = table_view{{rhs}};
+  auto table_comparator = cudf::detail::row::equality::two_table_comparator{tlhs, trhs, stream};
 
   auto outd = column_device_view::create(out, stream);
   auto optional_iter =
