+++ cpp/src/groupby/hash/helpers.cuh
@@ -16,8 +16,8 @@
 #pragma once
 
 #include <cudf/detail/cuco_helpers.hpp>
+#include <cudf/detail/row_operator/row_operators.cuh>
 #include <cudf/detail/utilities/integer_utils.hpp>
-#include <cudf/table/experimental/row_operators.cuh>
 #include <cudf/types.hpp>
 
 #include <cuco/static_set.cuh>
@@ -52,9 +52,8 @@ using shmem_extent_t =
 CUDF_HOST_DEVICE auto constexpr valid_extent =
   cuco::make_valid_extent<GROUPBY_CG_SIZE, GROUPBY_BUCKET_SIZE>(shmem_extent_t{});
 
-using row_hash_t =
-  cudf::experimental::row::hash::device_row_hasher<cudf::hashing::detail::default_hash,
-                                                   cudf::nullate::DYNAMIC>;
+using row_hash_t = cudf::detail::row::hash::device_row_hasher<cudf::hashing::detail::default_hash,
+                                                              cudf::nullate::DYNAMIC>;
 
 /// Adapter to cudf row hasher with caching support.
 class row_hasher_with_cache_t {
@@ -78,15 +77,15 @@ class row_hasher_with_cache_t {
 /// Probing scheme type used by groupby hash table
 using probing_scheme_t = cuco::linear_probing<GROUPBY_CG_SIZE, row_hasher_with_cache_t>;
 
-using row_comparator_t = cudf::experimental::row::equality::device_row_comparator<
+using row_comparator_t = cudf::detail::row::equality::device_row_comparator<
   false,
   cudf::nullate::DYNAMIC,
-  cudf::experimental::row::equality::nan_equal_physical_equality_comparator>;
+  cudf::detail::row::equality::nan_equal_physical_equality_comparator>;
 
-using nullable_row_comparator_t = cudf::experimental::row::equality::device_row_comparator<
+using nullable_row_comparator_t = cudf::detail::row::equality::device_row_comparator<
   true,
   cudf::nullate::DYNAMIC,
-  cudf::experimental::row::equality::nan_equal_physical_equality_comparator>;
+  cudf::detail::row::equality::nan_equal_physical_equality_comparator>;
 
 using global_set_t = cuco::static_set<cudf::size_type,
                                       cuco::extent<int64_t>,
