+++ cpp/src/io/orc/dict_enc.cu
@@ -17,10 +17,10 @@
 #include "orc_gpu.hpp"
 
 #include <cudf/detail/offsets_iterator.cuh>
+#include <cudf/detail/row_operator/row_operators.cuh>
 #include <cudf/detail/utilities/integer_utils.hpp>
 #include <cudf/hashing/detail/murmurhash3_x86_32.cuh>
 #include <cudf/io/orc_types.hpp>
-#include <cudf/table/experimental/row_operators.cuh>
 
 #include <rmm/cuda_stream_view.hpp>
 
@@ -82,7 +82,7 @@ struct equality_functor {
   __device__ bool operator()(size_type lhs_idx, size_type rhs_idx) const
   {
     // We don't call this for nulls so this is fine
-    auto const equal = cudf::experimental::row::equality::nan_equal_physical_equality_comparator{};
+    auto const equal = cudf::detail::row::equality::nan_equal_physical_equality_comparator{};
     return equal(col.element<string_view>(lhs_idx), col.element<string_view>(rhs_idx));
   }
 };
