+++ cpp/src/join/distinct_hash_join.cu
@@ -18,10 +18,10 @@
 #include <cudf/detail/cuco_helpers.hpp>
 #include <cudf/detail/join/distinct_hash_join.cuh>
 #include <cudf/detail/nvtx/ranges.hpp>
+#include <cudf/detail/row_operator/primitive_row_operators.cuh>
+#include <cudf/detail/row_operator/row_operators.cuh>
 #include <cudf/detail/utilities/cuda.cuh>
 #include <cudf/join/distinct_hash_join.hpp>
-#include <cudf/table/experimental/row_operators.cuh>
-#include <cudf/table/primitive_row_operators.cuh>
 #include <cudf/table/table_view.hpp>
 #include <cudf/types.hpp>
 #include <cudf/utilities/memory_resource.hpp>
@@ -53,7 +53,7 @@ bool constexpr has_nulls = true;  ///< Always has nulls
  */
 template <typename T>
 class primitive_keys_fn {
-  using hasher = cudf::row::primitive::row_hasher<>;
+  using hasher = cudf::detail::row::primitive::row_hasher<>;
 
  public:
   CUDF_HOST_DEVICE constexpr primitive_keys_fn(hasher const& hash) : _hash{hash} {}
@@ -72,9 +72,8 @@ class primitive_keys_fn {
  */
 template <typename T>
 class build_keys_fn {
-  using hasher =
-    cudf::experimental::row::hash::device_row_hasher<cudf::hashing::detail::default_hash,
-                                                     cudf::nullate::DYNAMIC>;
+  using hasher = cudf::detail::row::hash::device_row_hasher<cudf::hashing::detail::default_hash,
+                                                            cudf::nullate::DYNAMIC>;
 
  public:
   CUDF_HOST_DEVICE constexpr build_keys_fn(hasher const& hash) : _hash{hash} {}
@@ -168,8 +167,7 @@ distinct_hash_join::distinct_hash_join(cudf::table_view const& build,
   : _has_nested_columns{cudf::has_nested_columns(build)},
     _nulls_equal{compare_nulls},
     _build{build},
-    _preprocessed_build{
-      cudf::experimental::row::equality::preprocessed_table::create(_build, stream)},
+    _preprocessed_build{cudf::detail::row::equality::preprocessed_table::create(_build, stream)},
     _hash_table{build.num_rows(),
                 load_factor,
                 cuco::empty_key{cuco::pair{std::numeric_limits<hash_value_type>::max(),
@@ -207,16 +205,16 @@ distinct_hash_join::distinct_hash_join(cudf::table_view const& build,
     }
   };
 
-  if (cudf::is_primitive_row_op_compatible(_build)) {
-    auto const d_hasher =
-      cudf::row::primitive::row_hasher{nullate::DYNAMIC{has_nulls}, this->_preprocessed_build};
+  if (cudf::detail::is_primitive_row_op_compatible(_build)) {
+    auto const d_hasher = cudf::detail::row::primitive::row_hasher{nullate::DYNAMIC{has_nulls},
+                                                                   this->_preprocessed_build};
 
     auto const iter = cudf::detail::make_counting_transform_iterator(
       0, primitive_keys_fn<rhs_index_type>{d_hasher});
 
     build_hash_table(iter);
   } else {
-    auto const row_hasher = experimental::row::hash::row_hasher{this->_preprocessed_build};
+    auto const row_hasher = detail::row::hash::row_hasher{this->_preprocessed_build};
     auto const d_hasher   = row_hasher.device_hasher(nullate::DYNAMIC{has_nulls});
 
     auto const iter =
@@ -251,12 +249,11 @@ distinct_hash_join::inner_join(cudf::table_view const& probe,
   auto const found_begin =
     thrust::make_transform_output_iterator(found_indices.begin(), output_fn{});
 
-  auto preprocessed_probe =
-    cudf::experimental::row::equality::preprocessed_table::create(probe, stream);
-  if (cudf::is_primitive_row_op_compatible(_build)) {
+  auto preprocessed_probe = cudf::detail::row::equality::preprocessed_table::create(probe, stream);
+  if (cudf::detail::is_primitive_row_op_compatible(_build)) {
     auto const d_hasher =
-      cudf::row::primitive::row_hasher{nullate::DYNAMIC{has_nulls}, preprocessed_probe};
-    auto const d_equal = cudf::row::primitive::row_equality_comparator{
+      cudf::detail::row::primitive::row_hasher{nullate::DYNAMIC{has_nulls}, preprocessed_probe};
+    auto const d_equal = cudf::detail::row::primitive::row_equality_comparator{
       nullate::DYNAMIC{has_nulls}, preprocessed_probe, _preprocessed_build, _nulls_equal};
     auto const iter = cudf::detail::make_counting_transform_iterator(
       0, primitive_keys_fn<lhs_index_type>{d_hasher});
@@ -270,10 +267,10 @@ distinct_hash_join::inner_join(cudf::table_view const& probe,
                                found_begin,
                                stream);
   } else {
-    auto const two_table_equal = cudf::experimental::row::equality::two_table_comparator(
-      preprocessed_probe, _preprocessed_build);
+    auto const two_table_equal =
+      cudf::detail::row::equality::two_table_comparator(preprocessed_probe, _preprocessed_build);
 
-    auto const probe_row_hasher = cudf::experimental::row::hash::row_hasher{preprocessed_probe};
+    auto const probe_row_hasher = cudf::detail::row::hash::row_hasher{preprocessed_probe};
     auto const d_probe_hasher   = probe_row_hasher.device_hasher(nullate::DYNAMIC{has_nulls});
     auto const iter             = cudf::detail::make_counting_transform_iterator(
       0, build_keys_fn<lhs_index_type>{d_probe_hasher});
@@ -346,13 +343,12 @@ std::unique_ptr<rmm::device_uvector<size_type>> distinct_hash_join::left_join(
   auto const output_begin =
     thrust::make_transform_output_iterator(build_indices->begin(), output_fn{});
 
-  auto preprocessed_probe =
-    cudf::experimental::row::equality::preprocessed_table::create(probe, stream);
+  auto preprocessed_probe = cudf::detail::row::equality::preprocessed_table::create(probe, stream);
 
-  if (cudf::is_primitive_row_op_compatible(_build)) {
+  if (cudf::detail::is_primitive_row_op_compatible(_build)) {
     auto const d_hasher =
-      cudf::row::primitive::row_hasher{nullate::DYNAMIC{has_nulls}, preprocessed_probe};
-    auto const d_equal = cudf::row::primitive::row_equality_comparator{
+      cudf::detail::row::primitive::row_hasher{nullate::DYNAMIC{has_nulls}, preprocessed_probe};
+    auto const d_equal = cudf::detail::row::primitive::row_equality_comparator{
       nullate::DYNAMIC{has_nulls}, preprocessed_probe, _preprocessed_build, _nulls_equal};
 
     auto const iter = cudf::detail::make_counting_transform_iterator(
@@ -374,10 +370,10 @@ std::unique_ptr<rmm::device_uvector<size_type>> distinct_hash_join::left_join(
                    build_indices->end(),
                    JoinNoneValue);
     } else {
-      auto const two_table_equal = cudf::experimental::row::equality::two_table_comparator(
-        preprocessed_probe, _preprocessed_build);
+      auto const two_table_equal =
+        cudf::detail::row::equality::two_table_comparator(preprocessed_probe, _preprocessed_build);
 
-      auto const probe_row_hasher = cudf::experimental::row::hash::row_hasher{preprocessed_probe};
+      auto const probe_row_hasher = cudf::detail::row::hash::row_hasher{preprocessed_probe};
       auto const d_probe_hasher   = probe_row_hasher.device_hasher(nullate::DYNAMIC{has_nulls});
       auto const iter             = cudf::detail::make_counting_transform_iterator(
         0, build_keys_fn<lhs_index_type>{d_probe_hasher});
