+++ cpp/src/reductions/segmented/nunique.cu
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2024, NVIDIA CORPORATION.
+ * Copyright (c) 2023-2025, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,9 +19,9 @@
 #include <cudf/column/column_device_view.cuh>
 #include <cudf/column/column_factories.hpp>
 #include <cudf/detail/labeling/label_segments.cuh>
+#include <cudf/detail/row_operator/row_operators.cuh>
 #include <cudf/reduction/detail/segmented_reduction.cuh>
 #include <cudf/reduction/detail/segmented_reduction_functions.hpp>
-#include <cudf/table/experimental/row_operators.cuh>
 #include <cudf/table/table_view.hpp>
 #include <cudf/types.hpp>
 #include <cudf/utilities/memory_resource.hpp>
@@ -63,9 +63,8 @@ std::unique_ptr<cudf::column> segmented_nunique(column_view const& col,
 
   // compute the unique identifiers within each segment
   auto const identifiers = [&] {
-    auto const d_col = column_device_view::create(col, stream);
-    auto const comparator =
-      cudf::experimental::row::equality::self_comparator{table_view({col}), stream};
+    auto const d_col      = column_device_view::create(col, stream);
+    auto const comparator = cudf::detail::row::equality::self_comparator{table_view({col}), stream};
     auto const row_equal =
       comparator.equal_to<false>(cudf::nullate::DYNAMIC{col.has_nulls()}, null_equality::EQUAL);
