+++ cpp/src/row_operator/primitive_row_operators.cu
@@ -14,16 +14,16 @@
  * limitations under the License.
  */
 
-#include <cudf/table/primitive_row_operators.cuh>
+#include <cudf/detail/row_operator/primitive_row_operators.cuh>
 #include <cudf/types.hpp>
 #include <cudf/utilities/traits.hpp>
 
 #include <algorithm>
 
-namespace cudf {
+namespace cudf::detail {
 bool is_primitive_row_op_compatible(cudf::table_view const& table)
 {
   return std::all_of(
     table.begin(), table.end(), [](auto const& col) { return cudf::is_numeric(col.type()); });
 }
-}  // namespace cudf
+}  // namespace cudf::detail
