+++ cpp/src/search/contains_table.cu
@@ -16,9 +16,9 @@
 
 #include "contains_table_impl.cuh"
 
+#include <cudf/detail/row_operator/primitive_row_operators.cuh>
+#include <cudf/detail/row_operator/row_operators.cuh>
 #include <cudf/detail/search.hpp>
-#include <cudf/table/experimental/row_operators.cuh>
-#include <cudf/table/primitive_row_operators.cuh>
 #include <cudf/table/table_view.hpp>
 #include <cudf/types.hpp>
 #include <cudf/utilities/traits.hpp>
@@ -47,9 +47,9 @@ rmm::device_uvector<bool> contains(table_view const& haystack,
   auto const has_any_nulls      = haystack_has_nulls || needles_has_nulls;
 
   auto const preprocessed_needles =
-    cudf::experimental::row::equality::preprocessed_table::create(needles, stream);
+    cudf::detail::row::equality::preprocessed_table::create(needles, stream);
   auto const preprocessed_haystack =
-    cudf::experimental::row::equality::preprocessed_table::create(haystack, stream);
+    cudf::detail::row::equality::preprocessed_table::create(haystack, stream);
 
   // The output vector.
   auto contained = rmm::device_uvector<bool>(needles.num_rows(), stream, mr);
@@ -60,15 +60,15 @@ rmm::device_uvector<bool> contains(table_view const& haystack,
     std::any_of(haystack.begin(), haystack.end(), [](auto const& col) {
       return cudf::is_floating_point(col.type());
     });
-  if (cudf::is_primitive_row_op_compatible(haystack) && !has_floating_point) {
-    auto const d_haystack_hasher =
-      cudf::row::primitive::row_hasher{nullate::DYNAMIC{has_any_nulls}, preprocessed_haystack};
-    auto const d_needle_hasher =
-      cudf::row::primitive::row_hasher{nullate::DYNAMIC{has_any_nulls}, preprocessed_needles};
+  if (cudf::detail::is_primitive_row_op_compatible(haystack) && !has_floating_point) {
+    auto const d_haystack_hasher = cudf::detail::row::primitive::row_hasher{
+      nullate::DYNAMIC{has_any_nulls}, preprocessed_haystack};
+    auto const d_needle_hasher = cudf::detail::row::primitive::row_hasher{
+      nullate::DYNAMIC{has_any_nulls}, preprocessed_needles};
     auto const d_hasher     = hasher_adapter{d_haystack_hasher, d_needle_hasher};
-    auto const d_self_equal = cudf::row::primitive::row_equality_comparator{
+    auto const d_self_equal = cudf::detail::row::primitive::row_equality_comparator{
       nullate::DYNAMIC{has_any_nulls}, preprocessed_haystack, preprocessed_haystack, compare_nulls};
-    auto const d_two_table_equal = cudf::row::primitive::row_equality_comparator{
+    auto const d_two_table_equal = cudf::detail::row::primitive::row_equality_comparator{
       nullate::DYNAMIC{has_any_nulls}, preprocessed_needles, preprocessed_haystack, compare_nulls};
     auto const d_equal = comparator_adapter{d_self_equal, d_two_table_equal};
     perform_contains(haystack,
@@ -81,15 +81,14 @@ rmm::device_uvector<bool> contains(table_view const& haystack,
                      contained,
                      stream);
   } else {
-    auto const haystack_hasher   = cudf::experimental::row::hash::row_hasher(preprocessed_haystack);
+    auto const haystack_hasher   = cudf::detail::row::hash::row_hasher(preprocessed_haystack);
     auto const d_haystack_hasher = haystack_hasher.device_hasher(nullate::DYNAMIC{has_any_nulls});
-    auto const needle_hasher     = cudf::experimental::row::hash::row_hasher(preprocessed_needles);
+    auto const needle_hasher     = cudf::detail::row::hash::row_hasher(preprocessed_needles);
     auto const d_needle_hasher   = needle_hasher.device_hasher(nullate::DYNAMIC{has_any_nulls});
     auto const d_hasher          = hasher_adapter{d_haystack_hasher, d_needle_hasher};
 
-    auto const self_equal =
-      cudf::experimental::row::equality::self_comparator(preprocessed_haystack);
-    auto const two_table_equal = cudf::experimental::row::equality::two_table_comparator(
+    auto const self_equal = cudf::detail::row::equality::self_comparator(preprocessed_haystack);
+    auto const two_table_equal = cudf::detail::row::equality::two_table_comparator(
       preprocessed_needles, preprocessed_haystack);
 
     if (cudf::detail::has_nested_columns(haystack)) {
