+++ cpp/src/search/contains_table_impl.cu
@@ -16,16 +16,16 @@
 
 #include "contains_table_impl.cuh"
 
+#include <cudf/detail/row_operator/row_operators.cuh>
 #include <cudf/hashing/detail/helper_functions.cuh>
-#include <cudf/table/experimental/row_operators.cuh>
 #include <cudf/utilities/memory_resource.hpp>
 
 #include <cuco/static_set.cuh>
 
 namespace cudf::detail {
 
-using cudf::experimental::row::lhs_index_type;
-using cudf::experimental::row::rhs_index_type;
+using cudf::detail::row::lhs_index_type;
+using cudf::detail::row::rhs_index_type;
 
 /**
  * @brief Build a row bitmask for the input table.
@@ -60,10 +60,9 @@ std::pair<rmm::device_buffer, bitmask_type const*> build_row_bitmask(table_view 
 
 // Explicit instantiations for non-nested types (HasNested=false)
 using hasher_adapter_t = hasher_adapter<
-  cudf::experimental::row::hash::device_row_hasher<cudf::hashing::detail::default_hash,
-                                                   nullate::DYNAMIC>,
-  cudf::experimental::row::hash::device_row_hasher<cudf::hashing::detail::default_hash,
-                                                   nullate::DYNAMIC>>;
+  cudf::detail::row::hash::device_row_hasher<cudf::hashing::detail::default_hash, nullate::DYNAMIC>,
+  cudf::detail::row::hash::device_row_hasher<cudf::hashing::detail::default_hash,
+                                             nullate::DYNAMIC>>;
 
 template void dispatch_nan_comparator<false, hasher_adapter_t>(
   table_view const& haystack,
@@ -73,20 +72,20 @@ template void dispatch_nan_comparator<false, hasher_adapter_t>(
   bool haystack_has_nulls,
   bool needles_has_nulls,
   bool has_any_nulls,
-  cudf::experimental::row::equality::self_comparator self_equal,
-  cudf::experimental::row::equality::two_table_comparator two_table_equal,
+  cudf::detail::row::equality::self_comparator self_equal,
+  cudf::detail::row::equality::two_table_comparator two_table_equal,
   hasher_adapter_t const& d_hasher,
   rmm::device_uvector<bool>& contained,
   rmm::cuda_stream_view stream);
 
 // For HasNested=false (non-nested columns) with nan_equal_comparator
-using nan_equal_self_comparator = cudf::experimental::row::equality::device_row_comparator<
+using nan_equal_self_comparator = cudf::detail::row::equality::device_row_comparator<
   false,
   cudf::nullate::DYNAMIC,
-  cudf::experimental::row::equality::nan_equal_physical_equality_comparator>;
+  cudf::detail::row::equality::nan_equal_physical_equality_comparator>;
 
 using nan_equal_two_table_comparator =
-  cudf::experimental::row::equality::strong_index_comparator_adapter<nan_equal_self_comparator>;
+  cudf::detail::row::equality::strong_index_comparator_adapter<nan_equal_self_comparator>;
 
 using nan_equal_comparator_adapter =
   comparator_adapter<nan_equal_self_comparator, nan_equal_two_table_comparator>;
@@ -102,13 +101,13 @@ template void perform_contains(table_view const& haystack,
                                rmm::cuda_stream_view stream);
 
 // For HasNested=false (non-nested columns) with nan_unequal_comparator
-using nan_unequal_self_comparator = cudf::experimental::row::equality::device_row_comparator<
+using nan_unequal_self_comparator = cudf::detail::row::equality::device_row_comparator<
   false,
   cudf::nullate::DYNAMIC,
-  cudf::experimental::row::equality::physical_equality_comparator>;
+  cudf::detail::row::equality::physical_equality_comparator>;
 
 using nan_unequal_two_table_comparator =
-  cudf::experimental::row::equality::strong_index_comparator_adapter<nan_unequal_self_comparator>;
+  cudf::detail::row::equality::strong_index_comparator_adapter<nan_unequal_self_comparator>;
 
 using nan_unequal_comparator_adapter =
   comparator_adapter<nan_unequal_self_comparator, nan_unequal_two_table_comparator>;
