+++ cpp/src/stream_compaction/distinct_helpers.cu
@@ -101,40 +101,40 @@ rmm::device_uvector<size_type> reduce_by_row(distinct_set_t<RowEqual>& set,
 }
 
 template rmm::device_uvector<size_type> reduce_by_row(
-  distinct_set_t<cudf::experimental::row::equality::device_row_comparator<
+  distinct_set_t<cudf::detail::row::equality::device_row_comparator<
     false,
     cudf::nullate::DYNAMIC,
-    cudf::experimental::row::equality::nan_equal_physical_equality_comparator>>& set,
+    cudf::detail::row::equality::nan_equal_physical_equality_comparator>>& set,
   size_type num_rows,
   duplicate_keep_option keep,
   rmm::cuda_stream_view stream,
   rmm::device_async_resource_ref mr);
 
 template rmm::device_uvector<size_type> reduce_by_row(
-  distinct_set_t<cudf::experimental::row::equality::device_row_comparator<
+  distinct_set_t<cudf::detail::row::equality::device_row_comparator<
     true,
     cudf::nullate::DYNAMIC,
-    cudf::experimental::row::equality::nan_equal_physical_equality_comparator>>& set,
+    cudf::detail::row::equality::nan_equal_physical_equality_comparator>>& set,
   size_type num_rows,
   duplicate_keep_option keep,
   rmm::cuda_stream_view stream,
   rmm::device_async_resource_ref mr);
 
 template rmm::device_uvector<size_type> reduce_by_row(
-  distinct_set_t<cudf::experimental::row::equality::device_row_comparator<
+  distinct_set_t<cudf::detail::row::equality::device_row_comparator<
     false,
     cudf::nullate::DYNAMIC,
-    cudf::experimental::row::equality::physical_equality_comparator>>& set,
+    cudf::detail::row::equality::physical_equality_comparator>>& set,
   size_type num_rows,
   duplicate_keep_option keep,
   rmm::cuda_stream_view stream,
   rmm::device_async_resource_ref mr);
 
 template rmm::device_uvector<size_type> reduce_by_row(
-  distinct_set_t<cudf::experimental::row::equality::device_row_comparator<
+  distinct_set_t<cudf::detail::row::equality::device_row_comparator<
     true,
     cudf::nullate::DYNAMIC,
-    cudf::experimental::row::equality::physical_equality_comparator>>& set,
+    cudf::detail::row::equality::physical_equality_comparator>>& set,
   size_type num_rows,
   duplicate_keep_option keep,
   rmm::cuda_stream_view stream,
